+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.6.2]
+
+### Added
+ 
+### Breaking Changes ###
+ 
+### Changed
+ - Minor DOCS.md updates.
+ - Updated minimum TLS version to v1.2 and enable v1.3 support.
+ - Migrate from pyopenssl to cryptography library and rework SSL encryption.
+  
+### Fixed
+ - Fix corrupt certificate throwing an unhandled exception.
+
+
 ## [1.6.1] - 2026-02-19
 
 ### Added